fix: move route to invoices
